Do not check for `globallyEnabled` on Feature Flags
This is something that was also different after the actual
implementation of the Feature Flag endpoint. Previously we may have
expected some disabled flags would still show up in the response but
with a `globallyEnabled` flag set to false. The current endpoint just
gives us a list of enabled flags.