Build and Upload Storybook in Github Workflows
To make Storybook easier to access for non-developers, we want to build
a static version of it that gets deployed to each environment. This
requires building a copy of Storybook alongside the main production
build and then uploading that static Storybook directory as well. This
is done by creating a new `build-storybook` step. Both Storybook and the
main production build are uploaded in the `upload-code` step.

Also update .gitignore to ignore the statically built Storybook
directory.

PER-9925: Build Storybook as part of web-app build process