Update .prettierignore
Update this to exclude the `index.html` files which include minimized
JavaScript that we do not want to format. Also exclude the vendored CSS
we're using so that it's easier to identify differences from the main
packages when we properly import them in the future.

PER-9990: Run prettier autoformatter on all eligible files