Fix extra call to `archive/accept` in Glam onboarding
The original onboarding flow was designed to not accept an archive
invitation until the end of the flow in case the user wanted to change
which archive they accept. In the new "Glam" onboarding flow, the user
can accept as many archives as they want in addition to creating an
archive. Accepting archive invites also happens as soon as the user
clicks accept on the first page, rather than at the end of onboarding.

Since the Glam onboarding flow is still built on top of the older
onboarding components, this final archive accepting stage still runs
despite the fact that the invitations have been accepted. This causes a
brief error message to appear before the user is directed to their
archive. Only run this acceptance logic during the legacy onboarding
flow so that this doesn't happen.

This error could have been prevented if the "glam" onboarding was built
separately from the previous onboarding flow. This has caused many of
the recent onboarding issues in both Glam and original flows.

PER-9971: Final onboarding screen triggers an error for invitation