Adjust sign in/sign up page to support announcements
We want to display a downtime banner soon, but the new "glam" auth page
does not support showing the banner. Well, it does, but it doesn't
account for it in the layout and some flexbox nightmares manifest and
ruin the whole layout of the page.

Adjust the page so that it can handle the announcement box without
disrupting the layout of the screen. The specific challenge in this is
making the rest of the page responsive to its presence in a way that
does not cause scrollbars to appear.

This reworks how the image appears on signup right now. Previously it
was an `<img>` tag. Now it is a background image again, this time with
an `aria-description` that describes the text instead of an `img` tag
with an `alt` attribute. This also means the image "viewport" can change
aspect ratios without affecting the image, which makes for a more
aesthetically pleasing layout in my opinion.

This may have inadvertently fixed issues with scrollbars appearing on
the sign in screen.

PER-9980: Announcement banner doesn't show up right on sign in screen