Add edge cases for Public Route generation
There are a few edge cases in the PublicRoutePipe that are not addressed
and so they cause some incorrect URLs to be generated.

First, make it so that public objects in the public root do not need to
feature their parent folder's archiveNbr and folder_linkId. This makes
URLs match those that are found in the actual public gallery.

Second, address issues when it comes to mismatches in a record/folder's
archiveNbr and the archiveNbr of the archive the item is contained in.
In theory these numbers should always match, and the original (and new
fallback) implementation utilizes this assumption. In practice, there
are records in the public gallery in production that have archiveNbr
mismatches. This is concerning and we should figure out what's causing
it on the back-end, but we also need to adjust the link generation code
to handle these existing records.

Use the `archiveArchiveNbr` on the record as the canonical source of
the parent archive's ArchiveNbr. That is what this field is for so we
should read it instead of trying to reverse-engineer it ourselves.
However, in the case of the errant record on production, this alone does
not fix URL generation since it's actually null. As a fallback, check
the parent folder's `archiveArchiveNbr` as this value is properly set in
the production bug we've found. Finally use the old algorithm if
everything else fails.

PER-9988: Ensure the public gallery links generate accurately