Set up dialogs to accept hash/fragment
Previously some dialogs had hardcoded logic to map specific query
parameters to specific tabs in these dialogs. Generalize this code so
that it applies to every tab in these dialogs and also use the route
fragment instead of query parameters. In general, URI fragments are used
to identify a specific section of a resource so semantically it makes
the most sense to use fragments when linking to specific sections of our
modals.

Ideally we should really reduce more of the duplicated code used in
these dialog components.

PER-9973: Users can get help with checklist items when clicked