Link to dialogs/help on checklist items
Create a pipe so that the HTML template can look up whether a task has a
defined link or not. Querying the pipe within the template multiple
times is awkward, but it is preferred to storing the data in component
properties.

PER-9973: Users can get help with checklist items when clicked