Adjust background image(s) for Legacy Lab promo
Update CSS and HTML for the auth screens to adjust how the background is
displayed for the Legacy Lab promo. Change out the background image as
well. This background uses a slicing technique to extend the vertical
stripes of the image without repeating the legacy lab circle[1].

Also use classes to isolate legacy lab specific background changes from
the defaults we want to use for normal images. That way, we can just
remove the references to classes in HTML without having to revert the
CSS back.

[1]: https://en.wikipedia.org/wiki/9-slice_scaling

Co-authored-by: Han Pham <han@permanent.org>