Revert deleting Legacy Lab Promo SCSS
We want to show this again in the future with text changes. We can
delete the Legacy Lab HTML, but we should keep the SCSS around to be
reused next time. The SCSS also acts as way to document what the
intended HTML structure looks like.

PER-10073: Switch login image