Do not send Mixpanel events if auth token is unset
The `event` endpoint expects a valid auth token. Trying to send an event
when not logged in causes an error to happen which refreshes the page.
On pages like the public gallery, this creates an endless loop of
refreshes. Do not send events if the token is missing, and if the token
is invalid, make sure to erase it so that the same token isn't tried
again.

PER-10086: Public gallery view event failes if not authenticated