Fetch tags on SearchService construction
For some reason this suddenly started causing a bug. I'm not sure how
that happened, but the SearchService might just not be constructed until
after the initial `getTags$()` event is published. Get the loaded tags
in the current tags service for the time being.

PER-10099: Keywords and custom metadata does not filter