WIP: Unlisted Shares API methods
This is where I left off with the unlisted shares API work.
Please note that the module should be renamed from `shares-unlisted` to
`share-links` and the `unlisted-shares-api` should be `share-links-api`.
This was the temporary name I was working on but I realize that this
functionality encompasses share links rather than just unlisted shares.

Also if you use this code please include me in a Co-authored-by tag or
my spirit will haunt this codebase forever