Update get record API call to use stela
We are beginning to migrate our new feature development to stela,
which means we need to begin using those endpoints here. Eventually
we may go even further and use the permanent SDK, but for now what
we are doing is shimming stela's output to look as though it came from
the PHP api.

By mapping the stela type to the v1 type we avoid needing to refactor
all of our components. Eventually we would want to refactor these
components to use a more streamlined data type, but that is a larger
lift than we want for now.

Using stela endpoints will allow us to benefit from new authentication
methods (e.g. unlisted share tokens) in the near future.

Issue #683 Support unlisted shares

Co-authored-by: crisnicandrei <62384997+crisnicandrei@users.noreply.github.com>