Upgrade to Angular 19
This is one step along the path of our ultimate update to Angular 20.

The contents of this commit came from following the Angular update guide
from 18 to 19[1]

In particular, running

```
ng update @angular/core@19 @angular/cli@19
```

This was followed by updating a variety of third party packages to use
versions that work with Angular 19.

There are two extra things to note:

1) I was getting errors related to CountUpDirective and tests; adding
   this to the list of things not to mock fixed the issue.
2) Shepard does not seem to support cjs properly any longer.  There is
   a patch to their build, but this is a long term liability.

Issue #590 Update to Angular 20

[1] https://angular.dev/update-guide?v=18.0-19.0&l=3