Upgrade to ESLint 9
ESLint 8 is no longer being maintained, and so we need to move to v9.
This change involves a major shift in the way configuration is done.

Making this improvement appears to have enabled some additional rules,
so those are now overridden with new rule disables in a new rule block.
We will be re-enabling them one-by-one alongside any necessary code
changes in the future.

Issue #591 Upgrade to eslint 9