Use `love` lint rules
The rest of our projects use love rules and we want to use them here as
well.  As expected we are violating a lot of the rules, so for now we
disable anything we aren't already following which will let us re-enable
one by one alongside any necessary code improvements.

Issue #591 Upgrade to eslint 9