Use tabs
A few years ago Prettier formally adopted tabs as their default.

All of our other projects use tabs.

As we update our lint rules and move to ESLint 9 it makes sense to get
this out of the way.

This is a painful commit since it changes every line and will result in
some short term merge conflicts but the best time to do it is always
"now"