Run `tsc` on lint
This runs tsc and also makes some fixes to make sure tsc actually
passes.  It turns out our typescript config did not have `skipLibCheck`
enabled (which is a recommended setting) so tsc was checking the types
of everything in node_modules.  That's nice and all, but type issues in
there aren't our problem.

Issue #624 Add tsc to our CI