Remove `@typescript-eslint/explicit-member-accessibility` override
This was both disabled AND specifically configured, which is a tad
confusing.

We may want to enable this some day, but our current configs don't
appear to, and the point of the exercise now is to bring us in line with
those configured community recommendations.

Issue #628 Enable disable lint rules