Create `storybook` dependabot group
There is an eslint plugin for storybook which appears to be peer-coupled
with storybook in a way that caused a recent dependabot update to fail.

What's awkward is that I imagine it is also peer-coupled with eslint,
which means there may be times when it has to be updated along side
one or the other (or both) peer dependencies.

For now I'll create a storybook group; I'm not sure how Dependabot will
handle situations where a package fits two groups, but this seems like
a fine way to find out.