Apply prettier consistently
Prettier is actually something we want to use for yml and json files, as
well as basically anything else that we check in as our own code.

Prettier also already ignores anything in `.gitignore` so we don't need
to duplicate ignored and generated files.

I kept the vendor stylesheets with the understanding that those are
ultimately not being maintained / modified by our team.

Issue #674 Apply prettier to more files