Fix billing payment processing
As part of a recent linting improvement we saw that BaseVO was doing
some implicit magic related to assigning constructor values to object
values. In an attempt to make that magic more explicit for
BillingPaymentVO I failed to account for the fact that some of the
object attributes (currency and refTableToIncrease) had a default value,
and that value was being replaced with `undefined` in the event that the
constructor wasn't being passed an explicit value (which generally it
wasn't).

This fixes that, and uses the default value if there is no override.

As an aside, ideally a type checker would have noticed that an optional
value was being unsafely assigned to an attribute that should not
contain undefined. Room for improvement there!