PER-10298: Change link type
With this change, the link type dropdown is visible. Changing it sends
an "update" call to the API that includes the correct value for link
type ('public' or 'private'). Next step is to update the actual API
call to point to the stela endpoint /share-links instead of the legacy
API.