Refine `eslint` dependabot group
The `eslint-plugin-storybook` seems to be updated closely with storybook
itself which is why we explicitly listed it as part of the `storybook`
group.

A recent `eslint` dependabot PR included it anyway, which goes against
dependabot's documentation (it should have been part of the storybook
group based on our current configuration).

This explicitly excludes it from `eslint` which I'm hoping fixes the
issue.  We'll find out the next time there's a version bump.