PER-10307: Show thumbnails correctly for public folders
Generally we don't use thumbnails on folders, except in the public
gallery. The stela shim uses different punctuation for thumbnail urls
than the old API. Account for either fieldname in the thumbnail
interface.

This is spreading the ambiguity to a new location, so it might be
preferable to use the thumbURL capitalization scheme in the original shim.