Add and delete share links with new API
Note that the change in the httpv2 service is likely over-broad. It
looks like the right way to do this is to use `{ observe: 'response'
}` instead of `{ observe: 'body' }` and then check the response
status, but I haven't found the right place to make that change.

In any case, this *does* make the promise resolve and the deletion
cycle now completes.