Handle empty response from deletion
I am concerned that this is an over-broad change. It looks like the
right way to do this is to use `{ observe: 'response' }` instead of `{
observe: 'body' }` and then check the response status, but I haven't
found the right place to make that change. Maybe @aasandei-vsp knows?

In any case, this *does* make the promise resolve and the deletion
cycle now completes.