Fix flaky test in data service
The method fetchFullItems in the data service chains multiple promises,
which normally wouldn't be a problem. But in this case, we do not get
the modified records at the end of chaining the promises, we actually
just modify the reference to the records list we sent as an argument.
This implementation yelds unpredictable results, so that is why
adding a timeout is the solution here, to make sure that the records
have been assigned the new value before checking.

Issue PER-10319 Fix flaky test in data service