Change private members of the file-list component into public
The properties router and isUnlistedShare were private members of the file-list
component, but they are used in the template, so they needed to be made public.
The code works, as it's javascript, but taking into account that we are using
Typescript, we should be using the correct modifiers.

Issue PER-10314 Implement folder in folder share preview for unlisted shares