Make sticky footer remain visible when the modal disappears
In order for the sticky footer to remain on the screen, when the modal disappears,
the listener had to be removed, because it specifically closes the component when
the user clicks outside of it. This seems to be more like legacy code that has been
used in the past, because the footer component is only used in the share preview and
the listener produces the bug.

Issue: PER-10316 Create account sticky footer disappears