Fix public gallery link for records
The public gallery links were not generating correctly, because the
endpoint for getting the record has changed and the property used to
create it, ParentFolderVOs, is not present anymore.
In order to create the record links, parent properties present on the
record object can be used.
One exception is, the link for records that are in the root folder
will have 2 extra parameters, but they will work.

Issue: PER-10348 Get public gallery link not working for individual records