Add fallback for stela record calls
Calls for records on stela can be made with two types of authentication:
- auth token
- share token
When making a call with the share token, if that record is not unrestricted,
the endpoint will return an empty list. When this happens, if the user is
logged in, we try to make the call for the record again, using the auth token,
because maybe the user does have access to it and we can still get it.

Issue: PER-10357 Fallback for share token