Fix tags not appearing in share sidebar
In the edit tags component, the item tags were mapped and filtered
to only match tags that were present on the current archive. This
seems to be redundant for items that are in the private workspace,
because the tags on the item and the tags on the archive have the
same type TagVOData.
However, the mapping and filtering would eliminate all tags for the
shared items, because none of them would be present in the archive's
tags.
In conclusion, eliminating the mapping/filtering would not affect the
private workspace and will fix the shared workspace tags.

Issue: PER-10365 Keywords and metadata should show in the sidebar for Share Files section