Create a stub folder when previewing an unlisted share record
When previewing a record as an unlisted share, the folder cannot
be retrieved because the logged in user does not have rights to
view it. This is why we need to create a mock folder that will be
saved in the data service and be used accross the components from
the share flow.

Issue: PER-10371 Unlisted share record preview broken