Make selected value non-clickable
A native HTML select cannot be triggered programatically to open,
because this would post a security risk, so clicking on the selected
value that is not part of the select element won't make the options
dropdown open. In order to make the selected value look like part
of the whole select flow, an absolute positioning is needed, which
was already implemented. The issue was that the element would still
be clickable, so cancelling the pointer events will make the click
trigger go through it, directly to the select element.

Issue: PER-10229 Archive-member-invite, access-role-dropdown, click area buggy