Use thumbnails from record if available, when creating stub folder
When previewing an unlisted share record, we need a mock folder
for it to be stored in the data service and provide it to different
components from the share flow. When creating it, we should use
the thumbnails from the current record if that is available or
else use generated previews.

Issue: PER-10371 Unlisted share record preview broken