Add ReplayWeb support for web archive files
We have added backend recognition for `wacz` files, which have the
ability to be "replayed" by a tool called ReplayWeb.

This is currently (1) behind a feature flag and (2) served from a dev
domain name, which reflects the fact that it is not quite ready for
production until we ensure replayed web pages don't have access to
permanent.org cookie data.

Issue #863 Complete WebRecorder project

Co-authored-by: Cecilia Krum <cecilia@permanent.org>
Co-authored-by: Daniel Schultz <slifty@gmail.com>
Co-authored-by: Andreea Asandei <andreea.asandei@vspartners.us>